+++ report_gen/CMakeLists.txt
@@ -5,6 +5,5 @@ project(report_gen)
 
 add_subdirectory(report_gen_lib)
 add_subdirectory(report_gen_ar)
-#add_subdirectory(report_gen_stats)
-add_subdirectory(report_gen_tool)
-
+add_subdirectory(report_compiler_lib)
+add_subdirectory(report_compiler)
